PMD-396
Most dialogs in PME don't edit the live object in case the user clicks Cancel.  That way we can simply throw away the copy the user was editing. The editBusinessModel method correctly copied the connection from the temp business model to the permanent business model; the newBusinessModel method incorrectly was not doing this for the connection.

Fix: Added the setConnection call in the newBusinessModel method.

git-svn-id: svn://source.pentaho.org/svnroot/pentaho-metadata-editor/trunk@32114 5d404267-db13-0410-8476-b1f6b398e63a